fix: 중복된 contextLoads 로 인한 테스트 실패 해결
