+++ src/test/java/roomescape/RoomescapeApplicationTest.java
@@ -1,13 +1,116 @@
 package roomescape;
 
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.hamcrest.Matchers.is;
 
-@SpringBootTest
-class RoomescapeApplicationTest {
+import io.restassured.RestAssured;
+import io.restassured.http.ContentType;
+import java.lang.reflect.Field;
+import java.sql.Connection;
+import java.sql.SQLException;
+import java.util.HashMap;
+import java.util.Map;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.test.annotation.DirtiesContext;
+import roomescape.controller.ReservationController;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
+public class RoomescapeApplicationTest {
+
+        @Autowired
+        private JdbcTemplate jdbcTemplate;
+
+        @Autowired
+        private ReservationController reservationController;
 
         @Test
-        void contextLoads() {
+        void 예약_조회() {
+                RestAssured.given().log().all()
+                        .when().get("/reservations")
+                        .then().log().all()
+                        .statusCode(200)
+                        .body("size()", is(0)); // 아직 생성 요청이 없으니 0개
         }
 
+        @Test
+        void 데이터베이스_연동() {
+                try (Connection connection = jdbcTemplate.getDataSource().getConnection()) {
+                        assertThat(connection).isNotNull();
+                        assertThat(connection.getCatalog()).isEqualTo("DATABASE");
+                        assertThat(connection.getMetaData().getTables(null, null, "RESERVATION", null).next()).isTrue();
+                } catch (SQLException e) {
+                        throw new RuntimeException(e);
+                }
+        }
+
+        @Test
+        void 시간_관리_API() {
+                Map<String, String> params = new HashMap<>();
+                params.put("startAt", "10:00");
+
+                RestAssured.given().log().all()
+                        .contentType(ContentType.JSON)
+                        .body(params)
+                        .when().post("/times")
+                        .then().log().all()
+                        .statusCode(200);
+
+                RestAssured.given().log().all()
+                        .when().get("/times")
+                        .then().log().all()
+                        .statusCode(200)
+                        .body("size()", is(1));
+
+                RestAssured.given().log().all()
+                        .when().delete("/times/1")
+                        .then().log().all()
+                        .statusCode(200);
+        }
+
+        @Test
+        void 예약과_시간_연결() {
+                Map<String, String> time = new HashMap<>();
+                time.put("startAt", "10:00");
+
+                RestAssured.given().log().all()
+                        .contentType(ContentType.JSON)
+                        .body(time)
+                        .when().post("/times");
+
+                Map<String, Object> reservation = new HashMap<>();
+                reservation.put("name", "브라운");
+                reservation.put("date", "2026-08-05");
+                reservation.put("timeId", 1);
+
+                RestAssured.given().log().all()
+                        .contentType(ContentType.JSON)
+                        .body(reservation)
+                        .when().post("/reservations")
+                        .then().log().all()
+                        .statusCode(200);
+
+                RestAssured.given().log().all()
+                        .when().get("/reservations")
+                        .then().log().all()
+                        .statusCode(200)
+                        .body("size()", is(1));
+        }
+
+        @Test
+        void 계층화_리팩터링() {
+                boolean isJdbcTemplateInjected = false;
+
+                for (Field field : reservationController.getClass().getDeclaredFields()) {
+                        if (field.getType().equals(JdbcTemplate.class)) {
+                                isJdbcTemplateInjected = true;
+                                break;
+                        }
+                }
+
+                assertThat(isJdbcTemplateInjected).isFalse();
+        }
 }
